Fix build for web
